requirements.txt
fastapi[all]
sqlalchemy
psycopg2
pydantic
python-dotenv
pytest